Fix gcc false positive pedantic error:
https://travis-ci.org/biojppm/c4stl/jobs/241012139

This error was happening with signed size types. gcc was seeing that when i
was negative, the branch for (m_capacity <= N) would be the one to be hit. As
a result, m_arr[i] would be returned and since i was negative and m_arr is an
array, an out-of-bounds access would be proved to occur.

This was fixed by adding a C4_LIKELY() annotation and adding an i>=0 to its
condition. With this fix, when i is negative an out-of-bounds access will
still occur, but this time in the pointer, which (for the compiler) may turn
out to be valid, so a warning (error with -Werror) is no longer issued. Of
course we do not ever want to hit a negative path, but that case will be
weeded out by the previous assert.